Clean up data structure tests
