+++ moldesign/_tests/test_data_structures.py
@@ -1,204 +1,12 @@
-""" Tests Molecule instances and other base classes they depend on
+""" Test various functionalities around data structure consistency
 """
-import collections
 import pickle
-import random
 
-import itertools
-import numpy as np
-import pytest
-
-import moldesign as mdt
 import moldesign.exceptions
 import moldesign.molecules.atomcollections
 import moldesign.utils.classes
-from moldesign import units as u
 
-from .helpers import get_data_path
-
-registered_types = {}
-
-
-def typedfixture(*types, **kwargs):
-    """This is a decorator that lets us associate fixtures with one or more arbitrary types.
-    We'll later use this type to determine what tests to run on the result"""
-
-    def fixture_wrapper(func):
-        for t in types:
-            registered_types.setdefault(t, []).append(func.__name__)
-        return pytest.fixture(**kwargs)(func)
-
-    return fixture_wrapper
-
-
-######################################
-# Test the basic data structures
-
-TESTDICT = collections.OrderedDict((('a', 'b'),
-                                    ('c', 3),
-                                    ('d', 'e'),
-                                    ('a', 1),
-                                    (3, 35)))
-
-
-@typedfixture('object')
-def dotdict():
-    dd = moldesign.utils.classes.DotDict(TESTDICT)
-    return dd
-
-
-@typedfixture('object')
-def ordered_dotdict():
-    dd = moldesign.utils.classes.OrderedDotDict(TESTDICT)
-    return dd
-
-
-@pytest.mark.parametrize('objkey', 'dotdict ordered_dotdict'.split())
-def test_dotdict(objkey, request):
-    dd = request.getfuncargvalue(objkey)
-
-    # basic lookups and functions
-    assert len(dd) == len(TESTDICT)
-    assert dd.a == TESTDICT['a']
-    assert dd.d == TESTDICT['d']
-    assert set(dd.keys()) == set(TESTDICT.keys())
-    assert set(dd.values()) == set(TESTDICT.values())
-
-    # __contains__ methods
-    for item in TESTDICT:
-        assert item in dd
-
-    # item deletion
-    del dd.d
-    assert 'd' not in dd
-    assert len(dd) == len(TESTDICT) - 1
-    del dd[3]
-    assert 3 not in dd
-
-    # equivalence of items and attrs
-    dd['k'] = 12345
-    assert getattr(dd, 'k') == 12345
-    setattr(dd, 'newkey', -42)
-    assert dd['newkey'] == -42
-
-
-def test_ordered_dotdict(ordered_dotdict):
-    assert ordered_dotdict.keys() == TESTDICT.keys()
-    assert ordered_dotdict.values() == TESTDICT.values()
-    assert ordered_dotdict.items() == TESTDICT.items()
-
-    for iterator_method in '__iter__ iterkeys itervalues iteritems'.split():
-        odd_iter = getattr(ordered_dotdict, iterator_method)()
-        test_iter = getattr(TESTDICT, iterator_method)()
-        for gotval, testval in itertools.izip_longest(odd_iter, test_iter, fillvalue=None):
-            assert gotval == testval, 'Iterator %s failed' % iterator_method
-
-
-# Some objects with units
-@typedfixture('object')
-def list_of_units(): return [1.0 * u.angstrom, 1.0 * u.nm, 1.0 * u.a0]
-
-
-@typedfixture('object', 'equality')
-def simple_unit_array(): return np.array([1.0, -2.0, 3.5]) * u.angstrom
-
-
-@typedfixture('object', 'equality')
-def unit_number(): return 391.23948 * u.ureg.kg * u.ang / u.alpha
-
-
-######################################
-# Test underlying elements
-@typedfixture('submolecule')
-def carbon_atom():
-    atom1 = mdt.Atom('C')
-    return atom1
-
-
-def test_carbon_atom(carbon_atom):
-    assert carbon_atom.symbol == 'C'
-    assert carbon_atom.mass == 12.0 * u.amu
-
-
-@typedfixture('submolecule')
-def carbon_copy(carbon_atom):
-    atoms = carbon_atom.copy()
-    return atoms
-
-
-def test_carbon_copy(carbon_copy, carbon_atom):
-    atom = carbon_copy
-    assert atom.symbol == carbon_atom.symbol
-    assert atom.mass == carbon_atom.mass
-    assert atom.bond_graph == {}
-
-
-######################################
-# Tests around hydrogen
-@typedfixture('molecule')
-def h2():
-    atom1 = mdt.Atom('H')
-    atom1.x = 0.5 * u.angstrom
-    atom2 = mdt.Atom(atnum=1)
-    atom2.position = [-0.5, 0.0, 0.0] * u.angstrom
-    h2 = mdt.Molecule([atom1, atom2], name='h2')
-    atom1.bond_to(atom2, 1)
-    return h2
-
-
-@typedfixture('molecule')
-def h2_harmonic():
-    mol = h2()
-    SPRING_CONSTANT = 1.0 * u.kcalpermol / (u.angstrom ** 2)
-    model = moldesign.models.HarmonicOscillator(k=SPRING_CONSTANT)
-    integrator = moldesign.integrators.VelocityVerlet(timestep=0.5*u.fs, frame_interval=30)
-    mol.set_energy_model(model)
-    mol.set_integrator(integrator)
-    return mol
-
-
-@typedfixture('submolecule')
-def h2_trajectory(h2_harmonic):
-    mol = h2_harmonic
-    mol.atoms[0].x = 1.0 * u.angstrom
-    mol.momenta *= 0.0
-    traj = mol.run(500)
-    return traj
-
-
-def test_h2_trajectory(h2_trajectory):
-    """
-    Check if the trajectory is the sine wave that we expect
-    """
-    traj = h2_trajectory
-    mol = traj.mol
-    k = mol.energy_model.params.k
-    period = 2*u.pi*np.sqrt(mol.atoms[0].mass/k)
-    for frame in traj.frames:
-        period_progress = (frame.time % period) / period
-        if period_progress < 0.1 or period_progress > 0.9:
-            # check for expected peaks of sine wave
-            assert frame.positions[0, 0] > 0.1 * u.angstrom
-        elif 0.4 < period_progress < 0.6:
-            # check for expected troughs of sine wave
-            assert frame.positions[0, 0] < -0.1 * u.angstrom
-
-
-@typedfixture('molecule')
-def h2_traj_tempmol(h2_trajectory):
-    return h2_trajectory._tempmol
-
-
-@typedfixture('molecule')
-def h2_harmonic_copy(h2_harmonic):
-    return mdt.Molecule(h2_harmonic)
-
-
-def test_h2_positions(h2):
-    atom1, atom2 = h2.atoms
-    assert (atom1.position == np.array([0.5, 0.0, 0.0]) * u.angstrom).all()
-    assert atom2.x == -0.5 * u.angstrom
-    assert atom1.distance(atom2) == 1.0 * u.angstrom
+from .test_objects import *
 
 
 def test_h2_protected_atom_arrays(h2):
@@ -234,21 +42,6 @@ def test_h2_array_link(h2):
     assert h2.atoms[1].py == 3.0*u.default.momentum
 
 
-def test_copy_breaks_link(h2):
-    h2copy = mdt.Molecule(h2)
-    h2.atoms[0].y = 4.0 * u.bohr
-    assert h2copy.atoms[0].y == 0.0 * u.angstrom
-    np.testing.assert_almost_equal(h2.positions[0,1].value_in(u.bohr),
-                                   4.0, 7)
-    assert h2copy.positions[0, 1] == 0.0 * u.bohr
-
-    h2copy.momenta[1, 0] = 2.0 * u.default.momentum
-    np.testing.assert_almost_equal(h2copy.atoms[1].px.value_in(u.default.momentum),
-                                   2.0, 7)
-    assert h2.momenta[1, 0] == 0.0 * u.default.momentum
-    assert h2.atoms[1].px == 0.0 * u.default.momentum
-
-
 def test_h2_harmonic_oscillator(h2_harmonic):
     mol = h2_harmonic
     atoms = h2_harmonic.atoms
@@ -265,36 +58,6 @@ def test_h2_harmonic_oscillator(h2_harmonic):
     force_units = u.kcalpermol/u.angstrom
     assert abs(f1 + f2).value_in(force_units) < 1e-14
     assert abs(f1 - 1.0*force_units).value_in(force_units) < 1e-14
-
-
-def test_h2_harmonic_copy_loses_simulation(h2_harmonic_copy, h2_harmonic):
-    mol = h2_harmonic_copy
-    assert mol.energy_model is mol.integrator is None
-    for name in 'num_atoms num_residues positions momenta'.split():
-        try:
-            assert getattr(h2_harmonic_copy, name) == getattr(h2_harmonic, name)
-        except ValueError:
-            assert (getattr(h2_harmonic_copy, name) == getattr(h2_harmonic, name)).all()
-
-    assert mol.atoms[0].bond_graph[mol.atoms[1]] == 1
-    assert mol.atoms[1].bond_graph[mol.atoms[0]] == 1
-
-
-def test_h2_calculation_caching(h2_harmonic):
-    h2 = h2_harmonic
-    h2.properties = mdt.MolecularProperties(h2)
-    true_energy = h2.calc_potential_energy()
-    assert 'potential_energy' in h2.properties
-    assert 'forces' in h2.properties
-    h2.potential_energy
-    h2.forces
-    h2.properties['potential_energy'] = 'banana'
-    assert h2.potential_energy == h2.calc_potential_energy() == 'banana'
-    props = h2.calculate()
-    assert props.potential_energy == h2.potential_energy == h2.calc_potential_energy() == 'banana'
-    props2 = h2.calculate(use_cache=False)
-    assert props2.potential_energy == h2.potential_energy == true_energy
-    assert h2.calc_potential_energy() == true_energy
 
 
 def test_h2_cache_flush(h2_harmonic):
@@ -326,134 +89,11 @@ def h2_properties_raises_not_calculated_yet(h2_harmonic):
         h2_harmonic.properties.potential_energy
 
 
-@typedfixture('submolecule')
-def copy_atoms_from_h2_harmonic(h2_harmonic):
-    atoms = h2_harmonic.atoms.copy()
-    return atoms
-
-@typedfixture('molecule')
-def h2_harmonic_thats_been_copied(h2_harmonic):
-    temp = mdt.Molecule(h2_harmonic)
-    return h2_harmonic
-
-
-def test_copying_doesnt_corrupt_original_h2_harmonic(h2_harmonic):
-    mol = h2_harmonic
-    integ = mol.integrator
-    model = mol.energy_model
-    residue = mol.residues[0]
-    chain = list(mol.chains)[0]
-    m2 = mdt.Molecule(mol)
-    assert integ is mol.integrator
-    assert model is mol.energy_model
-    assert len(mol.chains) == 1
-    assert len(mol.residues) == 1
-    assert residue == mol.residues[0]
-    assert chain == list(mol.chains)[0]
-
-
-def test_atoms_copied_from_h2_harmonic(copy_atoms_from_h2_harmonic, h2_harmonic):
-    atoms = copy_atoms_from_h2_harmonic
-    atom = atoms[0]
-    assert atom.molecule is None
-    assert atom.residue is not h2_harmonic.atoms[0].residue
-    assert atom.chain is not h2_harmonic.atoms[0].chain
-    assert atoms[0].residue is atoms[1].residue
-    assert atoms[0].chain is atoms[1].chain
-
-
-@typedfixture('submolecule')
-def h2_harmonic_atoms(h2_harmonic):
-    return h2_harmonic.atoms
-
-
 def test_h2_traj_energies(h2_trajectory):
     traj = h2_trajectory
     assert (np.abs(traj.positions[0,0]) <= 1.0 * u.angstrom).all()
     assert (traj.potential_energy <= 4.0 * u.kcalpermol).all()
 
-
-def test_h2(h2):
-    mol = h2
-    assert mol.num_atoms == 2
-    assert mol.atoms[0].symbol == mol.atoms[1].symbol == 'H'
-
-
-######################################
-# Tests around PDB ID 3AID
-@typedfixture('molecule', scope='session')
-def pdb3aid():
-    mol = mdt.from_pdb('3AID')
-    return mol
-
-
-@typedfixture('submolecule')
-def ligand_residue_3aid(pdb3aid):
-    unknown = pdb3aid.chains['A'](type='unknown')
-    assert len(unknown) == 1
-    return unknown[0]
-
-
-@typedfixture('submolecule')
-def ligand_3aid_atoms(ligand_residue_3aid):
-    return ligand_residue_3aid.atoms
-
-
-@typedfixture('molecule')
-def ligand3aid(ligand_residue_3aid):
-    newmol = mdt.Molecule(ligand_residue_3aid)
-    return newmol
-
-
-@pytest.fixture
-def random_atoms_from_3aid(pdb3aid):
-    atoms = moldesign.molecules.atomcollections.AtomList(random.sample(pdb3aid.atoms, 10))
-    return atoms
-
-
-def test_3aid(pdb3aid):
-    mol = pdb3aid
-    assert len(mol.chains) == 2
-
-
-def test_3aid_ligand_search(pdb3aid):
-    mol = pdb3aid
-    unknown = mol.chains['A'](type='unknown')
-    proteina = mol.chains['A'](type='protein')
-    proteinb = mol.chains['B'](type='protein')
-    assert len(unknown) == 1
-    assert len(proteina) == len(proteinb) == 99
-
-
-def test_ligand3aid(ligand3aid):
-    mol = ligand3aid
-    assert len(mol.chains) == 1
-    assert len(mol.residues) == 1
-
-
-def test_nucleic_build(nucleic):
-    mol = nucleic
-    assert mol.num_chains == 2
-    assert mol.num_residues == 8
-    assert mol.chains[0] is mol.chains['A']
-    assert mol.chains[1] is mol.chains['B']
-    assert len(mol.chains[0].residues) == len(mol.chains[1].residues) == 4
-
-
-
-######################################
-# Tests around a piece of DNA
-@typedfixture('molecule', scope='session')
-def nucleic():
-    # ACTG.pdb contains a molecule generated using mdt.build_dna('ACTG')
-    mol = mdt.read(get_data_path('ACTG.pdb'))
-    return mol
-
-
-######################################
-# Consistency tests for groups of objects
-moldesign_objects = registered_types['molecule'] + registered_types['submolecule']
-all_objects = registered_types['object'] + moldesign_objects
 
 
 @pytest.mark.parametrize('molkey', registered_types['molecule'])
@@ -568,16 +208,4 @@ def test_pickled_equality(objkey, request):
         except ValueError:
             assert (y == obj).all()
 
-
-@pytest.mark.parametrize('fixture_key', ['h2_harmonic_atoms',
-                                         'ligand_3aid_atoms',
-                                         'random_atoms_from_3aid'])
-def test_copy_atoms(fixture_key, request):
-    oldatoms = request.getfuncargvalue(fixture_key)
-    atoms = oldatoms.copy()
-    for old, new in zip(oldatoms, atoms):
-        assert old.residue is not new.residue
-        assert old.residue.name == new.residue.name
-        assert old.chain is not new.chain
-        assert old.chain.name == new.chain.name
